refactor: consolidate frontend and backend into single project
