--- frontend/go.mod
@@ -1,8 +1,0 @@
-module github.com/dan-nicholls/danlovesto.run/frontend
-
-go 1.24.5
-
-require (
-	github.com/a-h/templ v0.3.937 // indirect
-	github.com/go-chi/chi/v5 v5.2.2 // indirect
-)
