+++ go.mod
@@ -1,8 +1,9 @@
-module github.com/dan-nicholls/danlovesto.run/backend
+module github.com/dan-nicholls/danlovesto.run
 
 go 1.24.3
 
 require (
+	github.com/a-h/templ v0.3.943 // indirect
 	github.com/dustin/go-humanize v1.0.1 // indirect
 	github.com/google/uuid v1.6.0 // indirect
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
