+++ internal/web/stats/heatmap.go
