+++ internal/web/ui/pages/home.templ
@@ -1,8 +1,8 @@
 package pages
 
 import (
-	"github.com/dan-nicholls/danlovesto.run/frontend/internal/ui/layouts"
-	"github.com/dan-nicholls/danlovesto.run/frontend/internal/api"
+	"github.com/dan-nicholls/danlovesto.run/internal/web/ui/layouts"
+	"github.com/dan-nicholls/danlovesto.run/internal/web/apiclient"
 )
 
 templ Home(info api.Info, stats api.Summary, hmData api.HeatmapData) {
